Move enrichment functions into dedicated utility class.
